backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
httpx==0.27.2
scikit-learn==1.5.1
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.9
google-generativeai==0.8.3
gunicorn==23.0.0
pydantic-settings==2.5.2
firebase-functions==0.1.0
